Collision implement based off the middle point of the Player
I no longer reverse the world positions in tilemap i reverse pTiles when it is made.

Odd issue when doing collision but those floor, ceil, and round functions we learned helped. round was the one to fix it.

Now i want to add the collision box for the player at its feet.

+++ Game2/Framework/Source/Objects/SpriteSheet.h
@@ -27,11 +27,11 @@ public:
     virtual ~SpriteSheet();
 
     SpriteInfo GetSpriteInfo(std::string name);
-    float GetSheetWidth();
+    int GetSheetWidth();
     std::vector<SpriteInfo>(SpriteObjects);
    
 protected:
-    float m_SheetWidth = 0.0f;
+    int m_SheetWidth = 0;
   
 };
